Add commands to only show quick fix window results
This is a proof of concept requesting a new feature. Occasionally I’d like to perform a search and only view the results generated to the quick fix window instead of also viewing them in the larger editable panel. Having this ability would replace the need to also install plugins like ack.vim, ag.vim or ferret which display their results in either a quick fix or location list window. This would give all file searching the same API rather than flip flopping back and forth.

This commit adds 3 commands CtrlSFLocationPrompt, CtrlSFVwordPathLocation, CtrlSFVwordExecLocation to compliment the existing commands and behavior. For the most part the code has been duplicated and is not very DRY with the rest of the functions. No documentation has been added.